ADD : 컴포넌트 파일에서 상수 데이터만 따로 관리하도록 파일 분리
